feat: add husky
