+++ src/App.module.css
@@ -1,0 +1,18 @@
+.body {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+  max-width: 1000px;
+  margin: 0 auto;
+}
+.container {
+}
+header {
+  height: 100px;
+}
+main {
+  flex: 1 1 auto;
+}
+footer {
+}
