+++ src/components/Table/styles.css
@@ -1,9 +1,6 @@
 table {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
   gap: 10px;
-  max-width: 900px;
+  width: 100%;
   margin: 0 auto;
   background-color: #f6f7f8;
   border-radius: 5px;
@@ -18,6 +15,7 @@ td {
   justify-content: center;
   width: 90px;
   align-items: center;
+  padding: 5px;
 }
 
 td:first-child,
