+++ src/main.tsx
@@ -2,8 +2,8 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App.tsx';
-import { ErrorBoundary } from './components/ErrorBundary/ErrorBoundery';
-import { FallBack } from './components/fallBack/fallBack.tsx';
+import { ErrorBoundary } from './components/ErrorBundary/index.tsx';
+import { FallBack } from './components/FallBack/index.tsx';
 
 createRoot(document.getElementById('root') as HTMLElement).render(
   <StrictMode>
